requirements.txt
pandas
numpy
django>=4.2
djangorestframework
